Remove monkey-patching from functorch
Monkey-patching of torch.* APIs on `import functorch` isn't good so we're
removing it now that functorch comes installed alongside PyTorch.

This involves:
- adding the functorch tensor printing mechanism to `torch/_tensor_str.py`
- adding the error check for functorch and Tensor.backward() to the
implementation of Tensor.backward()

Test Plan:
- wait for CI

[ghstack-poisoned]
